+++ os/linux/kernel/slot.cc
@@ -25,7 +25,7 @@ namespace os {
 // Access the value of some kind of private slot (by reference). This is an
 // instance of the requested slot, although many such instances might actually
 // exist.
-intptr_t Slot(SlotCategory category, int sub_category) {
+intptr_t Slot(SlotCategory category, size_t sub_category) {
   switch (category) {
     case SLOT_VIRTUAL_REGISTER:
       GRANARY_ASSERT(sub_category < arch::MAX_NUM_SPILL_SLOTS);
